App/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Django==3.2.7
django-redis-cache==2.1.0
django-rq==2.1.0
django-translation-manager==1.1.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Pillow==8.3.2
polib==1.1.1
psycopg2-binary==2.9.1
pythonping==1.1.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
rq==1.10.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
